requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
greenlet==3.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
packaging==23.2
pika==1.3.2
pika-stubs==0.1.3
pluggy==1.3.0
prompt-toolkit==3.0.39
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-celery==0.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
flower==1.2.0